hapus header yang dulu
